Fix a merge error in commit 28f08d3753eb10a1393a63e6c581d43aad9f93b9

+++ storage/innobase/btr/btr0scrub.cc
@@ -150,7 +150,7 @@ btr_scrub_lock_dict_func(ulint space_id, bool lock_to_close_table,
 				"WARNING: %s:%u waited " TIMETPF " seconds for"
 				" dict_sys lock, space: " ULINTPF
 				" lock_to_close_table: %d\n",
-				file, line, long(now - start), space_id,
+				file, line, now - start, space_id,
 				lock_to_close_table);
 
 			last = now;
